server/requirements.txt
Flask_Migrate==2.5.2
Flask_SQLAlchemy==2.4.1
SQLAlchemy==1.3.11
Flask==1.1.1
Flask_Script==2.0.6
python-dotenv
pymysql